pyproject.toml
[tool.poetry]
name = "quotes-sticker-bot"
version = "0.1.0"
description = ""
authors = ["NekrodNIK"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
